Create AndrewRichardA3Q1.js
Created a sprite for a pirate ship using a combination of various beginShape/vertices, ellipses, and text reading "HMS PROCRASTINATOR". Drew the ship along the coordinates of mouseX, 540, so that the ship follows the mouse on a flat plane. Placed a giant blue rectangle somewhat overlapping the ship to sell the idea of it being in the ocean. Changed the background to a pale orange and added a big yellow sunset for extra flavor.
https://editor.p5js.org/Furi0usGeorgeX/sketches/Tic1YRpLB